requirements.txt
azure-ai-textanalytics==5.3.0b2
azure-ai-translation-text==1.0.0b1
bs4==0.0.1
docx2txt==0.8
faiss-cpu==1.7.3
langchain==0.0.170
openai==0.27.2
pre-commit==2.21.0
pycryptodome==3.17
pypdf==3.5.1
python-dotenv==1.0.0
streamlit==1.23.0
tiktoken==0.3.0
